% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_patterns.R
\name{extract_patterns}
\alias{extract_patterns}
\title{Extract pattern from text data}
\usage{
extract_patterns(text)
}
\arguments{
\item{text}{The text data to use (including corpus)}
}
\value{
A list containing three patterns, named as in Details.
}
\description{
Extract specific patterns from text document.
}
\details{
Identified patterns from the data are
\enumerate{
\item Punctuation_Words: words containing punctuation marks
\item Digits: numbers (digits)
\item Uppercase_Words: words starting with uppercase
}
}
\examples{
data("acq", package = "tm")
patterns_entire = extract_patterns(acq)
patterns = lapply(acq, extract_patterns)
print(patterns$'110')
print(patterns_entire)
}
