% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_optimal_k.R
\name{find_optimal_k}
\alias{find_optimal_k}
\title{Find the optimal number of topics}
\usage{
find_optimal_k(corpus, seed_number = NA, weights = c(0.5, 0.5), topic_range)
}
\arguments{
\item{corpus}{The corpus document to use}

\item{seed_number}{set.seed for Gibbs sampling in FindTopicNumber function ; defaults to NA. Object of class "integer".}

\item{weights}{A weighted ratio for each "Griffiths2004", "CaoJuan2009" metric ; defaults to c(0.5, 0.5). A vector containing two numeric elements. Ideally, the sum of vector is 1.}

\item{topic_range}{The considered number of topic range. A vector of type "integer" or "double". Ideally, the results from functions such as seq are recommended.}
}
\value{
A scalar of type "integer": The number of optimal topics from the corpus document
}
\description{
Automatically find the optimal number of topics from corpus document.
}
\details{
The objective of the function is finding the optimal number of topics based on FindTopicsNumber function in ldatuning package. Combination of two metrics "Griffiths2004", "CaoJuan2009" from FindTopicsNumber as a weighted sum is considered.
Weights can be set differently depending on the user's subjective view on the importance of the two measurements. Since the function normalises the vector to be ratio, the sum of the vector for the weights does not necessarily have to be 1.
The range of desired k (topic_range) must be manually set by users.
}
\examples{
data("acq", package = "tm")
topic_range <- seq(from=3,to=18,by=3)
find_optimal_k(acq, topic_range = topic_range)
find_optimal_k(acq, seed_number = 123, topic_range = topic_range)
find_optimal_k(acq, weights = c(0.3, 0.7), topic_range = topic_range)
find_optimal_k(acq, seed_number = 123, weights = c(4, 5), topic_range = topic_range)
}
