% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_models.R
\name{topic_models}
\alias{topic_models}
\title{Implement topic models}
\usage{
topic_models(
  corpus,
  model = topicmodels::LDA,
  k,
  seed_number = NA,
  n_terms = 20
)
}
\arguments{
\item{corpus}{The corpus document to use}

\item{model}{A topic model to use ; defaults to LDA. Only LDA and CTM are supported}

\item{k}{The number of topics. A scalar of type "integer"}

\item{seed_number}{set.seed for VEM estimation ; defaults to NA. Object of class "integer"}

\item{n_terms}{The number of top words by topic ; defaults to 20. A scalar of type "integer"}
}
\value{
A list containing three elements, named as in Details.
}
\description{
Estimate a topic model based on topicmodels package using for the VEM algorithm and present key metrics by topic and document.
}
\details{
The function is designed for presenting metrics by topic or document after estimating LDA (default) and CTM from topicmodels package.

The three metrics from the model are
\enumerate{
\item Top_words_by_topic: Top n words by topic. n_terms argument is used.
\item Latent_topic_prob: Topic probability for each document.
\item Highest_prop_topic: The highest probability topic in each document.
}
}
\examples{
data("acq", package = "tm")
topic_models(acq, k=7)
topic_models(acq, model = topicmodels::CTM, k = 7, n_terms = 10)
}
